Switch support chat from Crisp to Help Scout (Beacon)
Replaces the Crisp widget with Help Scout's Beacon across the app:
- app/layout.tsx: swap the Crisp loader script for the Beacon snippet
  (init acb6362b-…).
- helpscout-beacon-identifier.tsx (replaces crisp-user-identifier.tsx):
  identify the signed-in provider via Beacon('identify', …) with the same
  fields (email, name, role, school district/site, user_id). No student
  data is sent.
- navbar "Help" button now calls Beacon('open') instead of $crisp.
- beacon.d.ts replaces crisp.d.ts for the window global.
- Privacy policy + docs/subprocessors.md + docs/data-inventory.md updated
  to disclose Help Scout instead of Crisp (US/AWS; DPA pending per SPE-170).

Same data footprint (provider PII only); subprocessor swap done before the
district DPA is signed to avoid an NDPA change-notification later.
