+++ app/privacy/page.tsx
@@ -103,7 +103,7 @@ export default function PrivacyPage() {
               <li><strong>Vercel:</strong> Application hosting and infrastructure</li>
               <li><strong>Sentry:</strong> Error monitoring (operational error data, configured to minimize personal information)</li>
               <li><strong>OpenAI and Anthropic:</strong> AI providers for optional AI-assisted features. Limited student data (such as initials and IEP goal text) may be shared with them to generate lesson and assessment content.</li>
-              <li><strong>Crisp:</strong> Help chat support (may collect provider account info such as name, email, IP address, and browser type; users should avoid sharing student data in chat messages)</li>
+              <li><strong>Help Scout:</strong> Help desk and chat support (may collect provider account info such as name, email, IP address, and browser type; users should avoid sharing student data in chat messages)</li>
             </ul>
 
             <h2 className="text-xl font-semibold text-gray-900 mt-8 mb-4">5. Data Security</h2>
