+++ app/types/beacon.d.ts
@@ -1,0 +1,7 @@
+declare global {
+  interface Window {
+    Beacon: (method: string, options?: unknown, data?: unknown) => void;
+  }
+}
+
+export {};
