--- app/types/crisp.d.ts
@@ -1,8 +1,0 @@
-declare global {
-  interface Window {
-    $crisp: any[];
-    CRISP_WEBSITE_ID: string;
-  }
-}
-
-export {};
