+++ docs/subprocessors.md
@@ -27,7 +27,7 @@ _Last reviewed: 2026-06-11._
 | **Supabase** | Primary backend: Postgres DB, Auth, Storage, Realtime. System of record for all student + provider data. | **Yes — system of record.** | `lib/supabase/*`; storage uploads across `app/api/**`; project ref `qkcruccytmmdajfavpgb` | Region fixed at project creation (not runtime-configurable) |
 | **Vercel** | Production hosting (Next.js app, API routes, cron). All request/response traffic and runtime logs transit Vercel compute. | **Yes — in transit + logs.** | Deploy target; `next.config.js` (Vercel cron monitors, `maxDuration`) | Function region configurable |
 | **Sentry** | Error monitoring (exceptions + source-mapped stack traces). Sentry Logs and Session Replay are **disabled**, `sendDefaultPii: false`, and logger `meta`/`context` are no longer forwarded (SPE-167). Emails are scrubbed via `beforeSend`. | **Incidental only** — operational error data, minimized per SPE-167. | `sentry.server.config.ts`, `sentry.edge.config.ts`, `instrumentation-client.ts`, `lib/monitoring/sentry-scrub.ts`; org `chickenscratch` / project `speddy`; US ingest | Org region (currently US) |
-| **Crisp** | Support/help chat widget (`client.crisp.chat`), loaded via a `<Script>` tag (not an npm package — missed by the original SDK-based audit). | **Provider PII** — pushes the signed-in provider's name, email, and session metadata to Crisp; **no student data by design** (a provider could paste it into a chat message). | `app/layout.tsx`, `app/components/crisp-user-identifier.tsx` | Vendor-managed |
+| **Help Scout** | Support help desk + in-app chat widget (Beacon, `beacon-v2.helpscout.net`), loaded via a `<Script>` tag (not an npm package). | **Provider PII** — pushes the signed-in provider's name, email, and session metadata (role, school district/site, user ID) to Beacon; **no student data by design** (a provider could paste it into a chat message). | `app/layout.tsx`, `app/components/helpscout-beacon-identifier.tsx` | US (AWS); DPA pending — SPE-170 |
 
 ## Planned — disclosed but NOT currently enabled
